share top-level type splitter
